refactor: Operation 클래스 분리하기
